gridly: Simplify public function interface
This commit changes the public interface of two functions in GridLY
public interface, namely `gridPutMusic` and `gridSetSegmentTemplate`.
Both these functions took a mandatory music argument, optionally
preceded by a context modifier to provide additional arguments. As per
the discussion in issue #104, this is somewhat confusing in the case of
cells using many optional arguments, since it is not immediately clear
that the music belongs to the cells.

To simplify this situation, the music argument is "pulled" inside the
context modifier, so the parentheses of the `with` block explicitly mark
the boundaries of the cell definition.

This is a breaking change. In order to ease the migration, two functions
are provided, namely `gridPutMusicDepr` and `gridSetSegmentTemplateDepr`
that use the old interface. Please refer to the "Migration guide" in
GridLY's README for more information.