example.ly: limit line length
Lines beyond 80 characters are difficult to display in common setups, so rather have more and shorter lines.
Of course, this requires to break a single \editionMod[List] command into multiple lines, so these should be surrounded by empty lines. But I think this will remain legible or even be easier to overview.